+++ src/app/_blocks/BlogpostContent/index.tsx
@@ -6,14 +6,13 @@ import { sanitizeAndAddChapters } from '../../_utilities/sanitizeAndAddChapters'
 import styles from './styles.module.css'
 
 export default function BlogpostContent({ post }: { post: Blogpost }) {
-  const { summary, content_html, featuredImage } = post
-  const sanitizedContent = sanitizeAndAddChapters(content_html)
+  const { summary, featuredImage } = post
+
 
   return (
     <div className={styles.container}>
       <FeaturedImage className={styles.featuredImage} src={featuredImage} />
       <div className={styles.summary}>{summary}</div>
-      <div className={styles.content} dangerouslySetInnerHTML={{ __html: sanitizedContent }} />
     </div>
   )
 }
